+++ app/lib/input/keyboard_controller.dart
@@ -7,7 +7,7 @@
 import 'dart:async';
 
 import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart' show FocusNode, KeyEventResult;
+import 'package:flutter/widgets.dart' show FocusNode;
 
 import 'tilt_controller.dart' show ControlCommand;